- Wouter Wijngaards changed csBitArrayHashKeyHandler::ComputeHash   in bitarray.h, int to size_t to remove signed/unsigned warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20634 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/bitarray.h
@@ -423,7 +423,7 @@ class csBitArrayHashKeyHandler
 public:
   static unsigned int ComputeHash (const csBitArray& key)
   {
-    const int uintCount = sizeof (csBitArray::store_type) / sizeof (uint);
+    const size_t uintCount = sizeof (csBitArray::store_type) / sizeof (uint);
     union
     {
       csBitArray::store_type store;
